service-parser/requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
python-dotenv==0.19.2